backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# RAG dependencies (should match RAG-v1 requirements)
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
langchain-aws==0.1.0
chromadb==0.4.20
sentence-transformers==2.2.2
huggingface-hub==0.19.4
boto3>=1.34.72
PyMuPDF==1.23.14
rapidfuzz==3.6.1
